# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.createSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  handlers.errorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          details: {}
          message:
            type: string
        type: object
    type: object
  handlers.listSubscriptionsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handlers.subscriptionResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  handlers.subscriptionResponse:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  handlers.totalResponse:
    properties:
      from:
        type: string
      service_name:
        type: string
      to:
        type: string
      total:
        type: integer
      user_id:
        type: string
    type: object
info:
  contact: {}
  description: REST service for subscriptions aggregation
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      parameters:
      - description: Filter by user UUID
        in: query
        name: user_id
        type: string
      - description: Filter by service name (ILIKE)
        in: query
        name: service_name
        type: string
      - description: Limit (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.listSubscriptionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      parameters:
      - description: Subscription
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.subscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Delete subscription by ID
      tags:
      - subscriptions
    get:
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.subscriptionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Subscription
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.subscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Update subscription by ID (full replace)
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      parameters:
      - description: From month (MM-YYYY)
        in: query
        name: from
        required: true
        type: string
      - description: To month (MM-YYYY)
        in: query
        name: to
        required: true
        type: string
      - description: Filter by user UUID
        in: query
        name: user_id
        type: string
      - description: Filter by service name (ILIKE)
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.totalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Calculate total cost of subscriptions for period
      tags:
      - subscriptions
swagger: "2.0"